Extract CM test helpers.
