+++ packages/testing/src/multitest/mod.rs
@@ -1,1 +1,2 @@
+pub mod helpers;
 pub mod modules;
